requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0.25
aiomysql>=0.2.0
alembic>=1.13.0

# ML/CV
torch>=2.1.0
torchvision>=0.16.0
opencv-python-headless>=4.9.0
easyocr>=1.7.0
numpy>=1.26.0
Pillow>=10.2.0

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Logging
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
factory-boy>=3.3.0

# Development
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
